python-django/requirements.txt
# Python packages required for main.py

django
django-ckeditor

opencv-python

python-vlc
